requirements_minimal.txt
# Minimal requirements for Django API to run
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3

# Database - alternative to mysqlclient
PyMySQL==1.1.0

# Authentication & Security
djangorestframework-simplejwt==5.3.0

# API Documentation
drf-spectacular==0.26.5

# Utilities
python-decouple==3.8
requests==2.31.0
python-dateutil==2.8.2
pytz==2023.3
